FIX:테스트 api와 인증 api를 controller 단에서 분리 #2
