+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/user/service/UserAuthService.java
@@ -7,5 +7,4 @@ import backend.spectrum.dguonoff.domain.user.dto.SignUpRequest;
 public interface UserAuthService {
     String login(LoginRequest dto);
     void signUp(SignUpRequest dto);
-    AllUserResponse getAllUser();
 }
