+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/user/service/UserService.java
@@ -1,0 +1,9 @@
+package backend.spectrum.dguonoff.domain.user.service;
+
+import backend.spectrum.dguonoff.domain.user.dto.AllUserResponse;
+import org.springframework.security.core.Authentication;
+
+public interface UserService {
+    AllUserResponse getAllUser();
+
+}
